requirements.txt
backports.tempfile
boto3
click
packaging >= 18.0
twine >= 1.12.1
